backend_python/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
requests==2.32.3
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
tweepy==4.14.0
requests-oauthlib>=1.2.0,<2
pyngrok==7.2.2
werkzeug==3.1.3
